Backend/requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.0.7
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
phonenumbers==8.13.45
pillow==10.4.0
pytz==2024.2
PyYAML==6.0.2
setuptools==74.1.2
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
